Implement new tab for lifestyle
