+++ flutter-frontend/lib/screens/lifestyle_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class LifestyleScreen extends StatelessWidget {
+  const LifestyleScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(backgroundColor: Colors.blueGrey);
+  }
+}
